# docs/openapi.yaml
openapi: 3.0.0
info:
  title: k8n API
  version: 0.1.0
  description: REST API and WebSocket contract between k8n frontend and backend
servers:
  - url: http://localhost:8080
paths:
  /api/cluster/contexts:
    get:
      summary: List available Kubernetes contexts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/cluster/connect:
    post:
      summary: Connect to a specified context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                context:
                  type: string
      responses:
        '200':
          description: Successfully connected
        '400':
          description: Connection failed
  /api/cluster/resources:
    get:
      summary: List live cluster resources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    kind: { type: string }
                    name: { type: string }
                    namespace: { type: string }
                    labels:
                      type: object
                      additionalProperties: { type: string }
                    status: { type: string }
                    uid: { type: string }
  /ws/status:
    get:
      summary: WebSocket for real-time status updates (Informers)
      responses:
        '101':
          description: Switching Protocols
  /api/graph/apply:
    post:
      summary: Apply graph resources to cluster
      requestBody:
        content:
          application/yaml:
            schema:
              type: string
      responses:
        '200':
          description: Applied
  /api/schema/{kind}:
    get:
      summary: Get OpenAPI V3 schema for a given K8s Kind
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Clean JSON schema for the resource